improve tasks management
