+++ spec/lib/organisations/destroy_multiple_spec.rb
@@ -1,4 +1,6 @@
-describe Organisations::DestroyMultiple, type: :service do
+require Rails.root.join("lib/organisations/destroy_multiple")
+
+describe Organisations::DestroyMultiple do
   subject do
     described_class.call(organisation_ids:)
   end
